Separate ProjectKey into a top-level class
